Adds checks for oh-my-zsh and related files

+++ initial_setup-mac.sh
@@ -5,6 +5,8 @@
 #
 # Created by Dan Jones
 
+ZSH_CUSTOM=~/.oh-my-zsh/custom
+
 unamestr=`uname`
 if [[ "$unamestr" == 'Darwin' ]]; then
   echo "[-] Mac detected, proceeding with install"
@@ -14,7 +16,7 @@ else
 fi
 
 which brew
-if [[ $? != 0 ]] ; then
+if [[ $? != 0 ]]; then
     # Install Homebrew
     echo "[!] Installing Homebrew to install other dependancies"
     /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
@@ -23,32 +25,53 @@ else
     brew update
 fi
 
-echo "[-] Install oh-my-zsh"
-sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
-echo "[-] Oh-my-zsh install complete"
+if [[ -d ~/.oh-my-zsh ]]; then
+  echo "[-] Oh-my-zsh is installed, skipping."
+else
+  echo "[-] Installing oh-my-zsh"
+  sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+  echo "[-] Oh-my-zsh install complete"
 
-#Install ansible
-echo "[-] Installing Ansible"
-brew install ansible git
-echo "[-] Ansible install complete!"
+  echo "[-] Installs zsh completion/syntax highlighting"
+  git clone https://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
+  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
+  echo "[-] Completed install of zsh completion/syntax highlighting"
+fi
 
-echo "[-] Installs oh-my-zsh theme"
-git clone https://github.com/denysdovhan/spaceship-prompt.git "$ZSH_CUSTOM/themes/spaceship-prompt" --depth=1
-ln -s "$ZSH_CUSTOM/themes/spaceship-prompt/spaceship.zsh-theme" "$ZSH_CUSTOM/themes/spaceship.zsh-theme"
-echo "[-] Theme install complete"
+if [[ ! -f $ZSH_CUSTOM/themes/spaceship.zsh-theme ]]; then
+  echo "[-] Installing oh-my-zsh theme"
+  git clone https://github.com/denysdovhan/spaceship-prompt.git "$ZSH_CUSTOM/themes/spaceship-prompt" --depth=1
+  ln -s "$ZSH_CUSTOM/themes/spaceship-prompt/spaceship.zsh-theme" "$ZSH_CUSTOM/themes/spaceship.zsh-theme"
+  echo "[-] Theme install complete"
+fi
 
-echo "[-] Installs zsh completion/syntax highlighting"
-git clone https://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
-echo "[-] Completed install of zsh completion/syntax highlighting"
+if [[ ! -d ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions ]]; then
+  echo "[-] Installs zsh completion"
+  git clone https://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
+  echo "[-] Completed install of zsh completion"
+fi
+
+if [[ ! -d ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting ]]; then
+  echo "[-] Installs zsh syntax highlighting"
+  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
+  echo "[-] Completed install of zsh syntax highlighting"
+fi
 
 echo "[-] Ensure that permissions are correct for Oh-my-zsh"
-sudo chmod -R 755 ~/.oh-my-zsh
+chmod -R 755 ~/.oh-my-zsh
 echo "Done."
 
-echo "[-] Installing Ansible mods needed for playbooks"
-ansible-galaxy collection install community.general
-echo "[-] Mod install complete"
+#Install ansible
+which ansible
+if [[ $? != 0 ]] ; then
+  echo "[-] Installing Ansible"
+  brew install ansible git
+  echo "[-] Ansible install complete!"
+
+  echo "[-] Installing Ansible mods needed for playbooks"
+  ansible-galaxy collection install community.general
+  echo "[-] Mod install complete"
+fi
 
 echo "[-] Running Mac Ansible playbook"
 ansible-playbook laptop-initial_mac.yml
@@ -59,10 +82,10 @@ echo "
 yubikey_ssh_enable () {
     ssh-add -l 1> /dev/null 2>/dev/null
 #    if [[ $? -eq 0 ]] {
-        echo "Enabling Yubikey for use with SSH. Sudo password needed to restart gpg-agent with ssh support..."
+        echo \"Enabling Yubikey for use with SSH. Sudo password needed to restart gpg-agent with ssh support...\"
         sudo killall gpg-agent
         sudo killall ssh-agent  
-        eval $( gpg-agent --daemon --enable-ssh-support )
+        eval \$( gpg-agent --daemon --enable-ssh-support )
         ssh-add -l
 #    }
 }
